Focus test on what is being tested.
